backend/requirements.txt
Flask
PyJWT
flask-cors
flask-sqlalchemy
werkzeug
psycopg2-binary
python-dotenv
pymongo
